pyproject.toml
[tool.poetry]
name = "platforma-rekomendacje"
version = "0.1.0"
description = ""
authors = ["zielinskik <zielinskik@invicta.pl>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
pandas = ">1.3.4"
numpy = ">1.21.6"
SQLAlchemy = "^1.4.27"
psycopg2 = "^2.9.2"
matplotlib = "^3.5.1"
nltk = "^3.6.5"
spacy = "^3.2.1"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
strsimpy = "^0.2.1"
Unidecode = "^1.3.2"
pyarrow = "^6.0.1"
gensim = "^4.1.2"
pyLDAvis = "^3.3.1"
lightgbm = "^3.3.2"
shap = "^0.40.0"
openpyxl = "^3.0.9"
stringscore = "^0.1.0"
jellyfish = "^0.9.0"
phonetics = "^1.0.5"
Fuzzy = "^1.2.2"
Boruta = "^0.3"
BorutaShap = "^1.0.16"
xgboost = "^1.5.2"
mlxtend = "^0.19.0"
apyori = "^1.1.2"
fpgrowth-py = "^1.0.0"
tensorflow = "<=2.4.1"
abydos = "^0.5.0"
shutup = "^0.2.0"
imblearn = "^0.0"
scikit-criteria = "^0.7"
torch = "^1.12.0"
Bottleneck = "^1.3.5"
tensorboardX = "^2.5.1"
kPOD = "^0.18"
netgraph = "^4.9.7"
ipympl = "^0.9.2"
neo4j = "^5.1.0"
ampligraph = "^1.4.0"
Janitor = "^0.1.1"
pydot = "^1.4.2"
graphviz = "^0.20.1"
dash = "^2.6.1"
dill = "^0.3.6"
impyute = "^0.0.8"


[tool.poetry.dev-dependencies]
ipykernel = "^6.6.0"
python-dotenv = "^0.19.2"
blackcellmagic = "^0.0.3"
plotly = "^5.4.0"
wordcloud = "^1.8.1"
seaborn = "^0.11.2"
umap-learn = "^0.5.2"
datashader = "^0.13.0"
scikit-image = "^0.19.1"
bokeh = "^2.4.2"
holoviews = "^1.14.7"

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
